Refactored style dictionary config
