--- src/tokens/filters/is-color.js
@@ -1,6 +1,0 @@
-module.exports = {
-  name: 'isColor',
-  matcher: function(prop) {
-    return prop.attributes.category === 'color';
-  }
-};
