--- src/tokens/filters/is-type-scale.js
@@ -1,6 +1,0 @@
-module.exports = {
-  name: 'isTypeScale',
-  matcher: function(prop) {
-    return prop.attributes.category === 'ts';
-  }
-};
