--- src/tokens/filters/is-width.js
@@ -1,6 +1,0 @@
-module.exports = {
-  name: 'isWidth',
-  matcher: function(prop) {
-    return prop.attributes.category === 'width';
-  }
-};
